Initial changes & JSON screen rename
