+++ apps/admin/frontend/src/screens/definition_screen.tsx
@@ -1,11 +1,10 @@
 import React, { useContext, useState } from 'react';
-import { Contest, getContestDistrictName } from '@votingworks/types';
 import styled from 'styled-components';
 
 import { format } from '@votingworks/utils';
 import { assert } from '@votingworks/basics';
 
-import { Button, LinkButton, Prose, Text } from '@votingworks/ui';
+import { Button, Font, H3, Icons, LinkButton, P } from '@votingworks/ui';
 import { AppContext } from '../contexts/app_context';
 
 import { routerPaths } from '../router_paths';
@@ -13,15 +12,10 @@ import { routerPaths } from '../router_paths';
 import { NavigationScreen } from '../components/navigation_screen';
 import { RemoveElectionModal } from '../components/remove_election_modal';
 
-const ButtonListItem = styled.span`
-  display: block;
-  margin-bottom: 0.25em;
+const ButtonList = styled.div`
+  display: flex;
+  gap: 0.5rem;
 `;
-
-interface ContestGroup {
-  name: string;
-  contests: Contest[];
-}
 
 export function DefinitionScreen(): JSX.Element {
   const { electionDefinition, configuredAt } = useContext(AppContext);
@@ -30,85 +24,42 @@ export function DefinitionScreen(): JSX.Element {
 
   const [isRemovingElection, setIsRemovingElection] = useState(false);
 
-  const electionsByDistrict = election.contests.reduce<ContestGroup[]>(
-    (groups, contest) => {
-      const contestDistrictName = getContestDistrictName(election, contest);
-      const existingIndex = groups.findIndex(
-        (group) => group.name === contestDistrictName
-      );
-      if (existingIndex >= 0) {
-        groups[existingIndex].contests.push(contest);
-      } else {
-        groups.push({
-          name: contestDistrictName,
-          contests: [contest],
-        });
-      }
-      return groups;
-    },
-    []
-  );
-
   return (
     <React.Fragment>
       <NavigationScreen title="Election Definition">
-        <Prose maxWidth={false}>
-          <Text small>
-            Configured with the current election at{' '}
-            <strong>
-              {format.localeLongDateAndTime(new Date(configuredAt))}
-            </strong>
-          </Text>
-          <h2>Election Metadata</h2>
-          <p>
-            title: <strong>{election.title}</strong>
-            <br />
-            date: <strong>{election.date}</strong>
-            <br />
-            county name: <strong>{election.county.name}</strong>
-            <br />
-            state: <strong>{election.state}</strong>
-            <br />
-            seal: <strong>{election.sealUrl || election.seal}</strong>
-            <br />
-          </p>
-          <h2>Contests</h2>
-          {electionsByDistrict.map((district) => (
-            <React.Fragment key={district.name}>
-              <h3>{district.name}</h3>
-              <p>
-                {district.contests.map((contest) => (
-                  <ButtonListItem key={contest.id}>
-                    <LinkButton
-                      small
-                      to={routerPaths.definitionContest({
-                        contestId: contest.id,
-                      })}
-                    >
-                      {contest.title}
-                    </LinkButton>
-                  </ButtonListItem>
-                ))}
-              </p>
-            </React.Fragment>
-          ))}
-          <h1>Advanced Features</h1>
-          <p>
-            <ButtonListItem>
-              <LinkButton to={routerPaths.definitionEditor}>
-                View Definition JSON
-              </LinkButton>
-            </ButtonListItem>
-            <ButtonListItem>
-              <Button
-                variant="danger"
-                onPress={() => setIsRemovingElection(true)}
-              >
-                Remove Election
-              </Button>
-            </ButtonListItem>
-          </p>
-        </Prose>
+        <P>
+          Configured with the current election at{' '}
+          <Font weight="bold">
+            {format.localeLongDateAndTime(new Date(configuredAt))}
+          </Font>
+        </P>
+        <H3 as="h2">Election Metadata</H3>
+        <P>
+          <Font weight="semiBold">Title:</Font> {election.title}
+        </P>
+        <P>
+          <Font weight="semiBold">Date:</Font>{' '}
+          {format.localeLongDateAndTime(new Date(election.date))}
+        </P>
+        <P>
+          <Font weight="semiBold">County Name:</Font> {election.county.name}
+        </P>
+        <P>
+          <Font weight="semiBold">State:</Font> {election.state}
+        </P>
+        <P>
+          <Font weight="semiBold">Seal:</Font>{' '}
+          {election.sealUrl || election.seal}
+        </P>
+        <H3 as="h2">Advanced Features</H3>
+        <ButtonList>
+          <LinkButton to={routerPaths.definitionViewer}>
+            View Definition JSON
+          </LinkButton>
+          <Button onPress={() => setIsRemovingElection(true)}>
+            <Icons.Delete /> Remove Election
+          </Button>
+        </ButtonList>
       </NavigationScreen>
       {isRemovingElection && (
         <RemoveElectionModal onClose={() => setIsRemovingElection(false)} />
